refactor: 💡 Split randSuperhero and randSuperheroName
✅ Closes: #134
